Fix blank output when resampling to grid from SPH probe
The information about WholeExtents for Structured Grids needs to be set
in the RequestInformation pass, otherwise (if done in RequestData)
it gets lost during passes through other internal filters prior to display
in ParaView (vtkPVPostFilter).

Make a best guess for the information if using the GUI widgets by taking the resolution
and setting it during Requestinformation and update it whenever new info is available.