Fix SetStep on reader and writer
Make a small change to H5PArt so that we can overwrite the Step
name used for time step grouping.

Reader and Writer both overwrite global STEP var before opening
the file, so they must not use different step names concurrently.
Will work as long as one opens and closes in turn with the other.